a.yanin/common: Create a new common module `loop`
After prolonged deliberation, I've decided to create a simple event loop
that would support async TCP socket operations.

The event loop design is directly inspired by libuv,
with a few minor differences.

- the extensive use of the `error` and `log` modules because I don't
  care about other people's use-cases

- fearless memory allocations (since I've kept avoiding them for some
  reason while having a machine with 16 GiB of RAM)

- `poll` instead of the clearly superior `epoll` because POSIX...
  at least it's not `select`, duh.

- the extensible handler interface
  - they'll be responsible for juggling callbacks

- a decoupled executor interface (now that I think of it, it looks
  really similar to Java's executors... that's scary.)

I'm still thinking where to put the I/O operations in this picture
so that they wouldn't be a royal pain.

The goal is to have an overwhelmingly large part of the codebase shared
among the 31st and 33rd labs; the only difference there being
a different choice of the executor (single-threaded vs thread pool).

And since I anticipate (and actively hope) that this will be my last
grand effort writing code in C, I'm going extravagant with my designs.
To the detriment of the poor people who'll have to review this.
To them I say this: I'm not sorry at all.